added simulated button to fix recursion issues
there is now a simulated button that will only be use by one selected firefly. this code is used exclusively for testing, because we still like the button for demo, but it proved impractical for testing. To use this code, upload it to every firefly with swarmStarter = false, except one firefly. this one firefly will have swarmStarter = true, enabling it to initiate the swarm after a short delay. confirmed it works on group of 6: 1/28/17